miniapp/server/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
psycopg2-binary